+++ src/input/r_matroska.cpp
@@ -2551,6 +2551,7 @@ kax_reader_c::identify() {
     info.add(mtx::id::uid,                  track->track_uid);
     info.add(mtx::id::codec_id,             track->codec_id);
     info.set(mtx::id::codec_private_length, track->private_size);
+    info.add(mtx::id::codec_delay,          track->codec_delay.to_ns(0));
     info.add(mtx::id::language,             track->language);
     info.add(mtx::id::track_name,           track->track_name);
     info.add(mtx::id::stereo_mode,          static_cast<int>(track->v_stereo_mode), static_cast<int>(stereo_mode_c::unspecified));